ci(db): snapshot database and startup script

+++ docker/scripts/container-wordpress-command.sh
@@ -4,9 +4,13 @@
 docker-entrypoint.sh apache2
 
 # The database needs to be ready, so netcat it, and also the wp binary must be downloaded and executable
-apt-get update && apt-get -y install netcat
+# Also we need the mysql-client so mysqldump can be used by WP-CLI
+which netcat || apt-get update
+which netcat || apt-get -y install netcat mysql-client
+which netcat && echo "Skip apt-get installations because they are already installed on this volume"
 while ! nc -z mysql 3306; do sleep 1; done;
 while [ ! -f wp ]; do sleep 1; done;
+echo "WP CLI is available now"
 sleep 1; # Just to be sure the file is completely moved here
 
 # Move the installed wp-cli to bin so we can execute it
@@ -17,30 +21,45 @@ mv /var/www/html/wp /usr/local/bin/
 ### --------- DO NOT CHANGE UNTIL HERE --------- ###
 ####################################################
 
-# Install wordpress itself
-wp --allow-root core install --path="/var/www/html" --url="http://localhost" --title="wpdev" --admin_user=wordpress --admin_password=wordpress --admin_email=admin@test.com
+# Run the following scripts only when WordPress is started at first time
+# Use always --allow-root because docker runs the service as root user
+if ! $(wp --allow-root core is-installed); then
+    # Install wordpress itself
+    wp --allow-root core install --path="/var/www/html" --url="http://localhost" --title="wpdev" --admin_user=wordpress --admin_password=wordpress --admin_email=admin@test.com
 
-# Config parameters
-wp --allow-root config set WP_DEBUG true --raw
-wp --allow-root config set SCRIPT_DEBUG true --raw
-wp --allow-root config set FS_METHOD direct # see https://git.io/fj4IK, https://git.io/fj4Ii
+    # Config parameters
+    wp --allow-root config set WP_DEBUG true --raw
+    wp --allow-root config set SCRIPT_DEBUG true --raw
+    wp --allow-root config set FS_METHOD direct # see https://git.io/fj4IK, https://git.io/fj4Ii
 
-# Activate this plugin
-wp --allow-root plugin activate wp-reactjs-starter
+    # Activate this plugin
+    wp --allow-root plugin activate wp-reactjs-starter
 
-# Delete unnecessery plugins
-rm -rf wp-content/plugins/akismet
-rm -rf wp-content/plugins/hello.php
+    # Delete unnecessery plugins
+    rm -rf wp-content/plugins/akismet
+    rm -rf wp-content/plugins/hello.php
+
+    # Import startup.sql script if available
+    test -f /scripts/startup.sql && wp --allow-root db import /scripts/startup.sql
+fi
 
 ####################################################
 ### --------- THIS MUST BE AT THE END --------- ###
 ####################################################
+
+# When CYPRESS test is active copy the e2e login plugin and activate it
+cp /scripts/e2e-tests-autologin-plugin.php wp-content/plugins/
+chown 33:33 wp-content/plugins/e2e-tests-autologin-plugin.php
+wp --allow-root plugin activate e2e-tests-autologin-plugin
 
 # Update only minor version
 wp --allow-root core update --minor
 
 # Change the rights so media / plugins / themes are possible to upload and install
 chown -R www-data:www-data /var/www/html/*
+
+# Create lock file so the host knows the WordPress installation is up and running
+touch /scripts/.env-wp.lock
 
 # Main CMD from https://git.io/fj4Fe
 apache2-foreground
